+++ contracts/oracle/IOracle.sol
@@ -6,12 +6,11 @@ import "../external/Decimal.sol";
 /// @title generic oracle interface for Fei Protocol
 /// @author Fei Protocol
 interface IOracle {
-	
     // ----------- Events -----------
 
-	event KillSwitchUpdate(bool _killSwitch);
+    event KillSwitchUpdate(bool _killSwitch);
 
-	event Update(uint _peg);
+    event Update(uint256 _peg);
 
     // ----------- State changing API -----------
 
@@ -30,13 +29,13 @@ interface IOracle {
     /// @notice read the oracle price
     /// @return oracle price
     /// @return true if price is valid
-    /// @dev price is to be denominated in USD per X where X can be ETH, etc. 
+    /// @dev price is to be denominated in USD per X where X can be ETH, etc.
     /// @dev Can be innacurate if outdated, need to call `isOutdated()` to check
     function read() external view returns (Decimal.D256 memory, bool);
 
     /// @notice determine if read value is stale
     /// @return true if read value is stale
-    function isOutdated() external view returns(bool);
+    function isOutdated() external view returns (bool);
 
     /// @notice the kill switch for the oracle feed
     /// @return true if kill switch engaged
