+++ contracts/router/IFeiRouter.sol
@@ -3,7 +3,6 @@ pragma solidity ^0.6.0;
 /// @title A Uniswap Router for FEI/ETH swaps
 /// @author Fei Protocol
 interface IFeiRouter {
-
     // ----------- state changing api -----------
 
     /// @notice buy FEI for ETH with some protections
@@ -12,11 +11,11 @@ interface IFeiRouter {
     /// @param to address to send FEI
     /// @param deadline block timestamp after which trade is invalid
     function buyFei(
-        uint minReward, 
-        uint amountOutMin, 
-        address to, 
-        uint deadline
-    ) external payable returns(uint amountOut);
+        uint256 minReward,
+        uint256 amountOutMin,
+        address to,
+        uint256 deadline
+    ) external payable returns (uint256 amountOut);
 
     /// @notice sell FEI for ETH with some protections
     /// @param maxPenalty maximum fei burn for purchasing
@@ -25,10 +24,10 @@ interface IFeiRouter {
     /// @param to address to send ETH
     /// @param deadline block timestamp after which trade is invalid
     function sellFei(
-        uint maxPenalty, 
-        uint amountIn, 
-        uint amountOutMin, 
-        address to, 
-        uint deadline
-    ) external returns(uint amountOut);
+        uint256 maxPenalty,
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address to,
+        uint256 deadline
+    ) external returns (uint256 amountOut);
 }
