+++ contracts/token/IIncentive.sol
@@ -5,18 +5,17 @@ pragma solidity ^0.6.2;
 /// @notice Called by FEI token contract when transferring with an incentivized address
 /// @dev should be appointed as a Minter or Burner as needed
 interface IIncentive {
+    // ----------- Fei only state changing api -----------
 
-	// ----------- Fei only state changing api -----------
-
-	/// @notice apply incentives on transfer
-	/// @param sender the sender address of the FEI
-	/// @param receiver the receiver address of the FEI
-	/// @param operator the operator (msg.sender) of the transfer
-	/// @param amount the amount of FEI transferred
+    /// @notice apply incentives on transfer
+    /// @param sender the sender address of the FEI
+    /// @param receiver the receiver address of the FEI
+    /// @param operator the operator (msg.sender) of the transfer
+    /// @param amount the amount of FEI transferred
     function incentivize(
-    	address sender, 
-    	address receiver, 
-    	address operator, 
-    	uint amount
+        address sender,
+        address receiver,
+        address operator,
+        uint256 amount
     ) external;
 }
